// src/types/theme.ts
export interface ThemeSettings {
  shopName: string;
  currency: string;
  darkMode: boolean;
  notifications: boolean;
  logoUrl: string;
  primaryColor: string;
  accentColor: string;
  secondaryColor: string;
  sidebarColor: string; // Nouvelle propriété pour la couleur de la barre latérale
  borderRadius: string;
  fontFamily: string;
}